+++ __tests__/hooks/useWebSpeechPonyfill.js
@@ -1,5 +1,4 @@
 import { timeouts } from '../constants.json';
-import executePromiseScript from '../setup/pageObjects/executePromiseScript.js';
 
 // selenium-webdriver API doc:
 // https://seleniumhq.github.io/selenium/docs/api/javascript/module/selenium-webdriver/index_exports_WebDriver.html
@@ -18,13 +17,16 @@ test('getter should return webSpeechPonyfill from props', async () => {
   ).resolves.toBeTruthy();
 });
 
-test('getter should return undefined if not set in props', async () => {
+test('getter should return speech synthesis bypass if not set in props', async () => {
   const { pageObjects } = await setupWebDriver();
 
-  // Although useWebSpeechPonyfill return [undefined], runHook/executeAsyncScript turned it into [null].
   await expect(
-    pageObjects.runHook('useWebSpeechPonyfill', [], webSpeechPonyfill => webSpeechPonyfill[0])
-  ).resolves.toBeNull();
+    pageObjects.runHook('useWebSpeechPonyfill', [], webSpeechPonyfill =>
+      Object.keys(webSpeechPonyfill[0])
+        .sort()
+        .join(', ')
+    )
+  ).resolves.toMatchInlineSnapshot(`"SpeechSynthesisUtterance, speechSynthesis"`);
 });
 
 test('setter should throw exception', async () => {