+++ packages/directlinespeech/__tests__/utilities/createTestHarness.js
@@ -1,6 +1,5 @@
-import createDeferred from 'p-defer';
-
 import createAdapters from '../../src/createAdapters';
+import createDeferred from '../../src/external/p-defer';
 import createQueuedArrayBufferAudioSource from './createQueuedArrayBufferAudioSource';
 import fetchSpeechCredentialsWithCache from './fetchSpeechCredentialsWithCache';
 import fetchSpeechData from './fetchSpeechData';