+++ packages/directlinespeech/src/createWebSpeechPonyfillFactory.js
@@ -2,8 +2,8 @@
 
 import { AudioConfig } from 'microsoft-cognitiveservices-speech-sdk';
 import { createSpeechRecognitionPonyfillFromRecognizer } from 'web-speech-cognitive-services/lib/SpeechServices/SpeechToText';
-import AbortController from 'abort-controller';
 
+import AbortController from './external/abort-controller';
 import createErrorEvent from './createErrorEvent';
 import createTaskQueue from './createTaskQueue';
 import EventTargetShim, { defineEventAttribute } from './external/event-target-shim';
@@ -21,6 +21,14 @@ export default function({
   recognizer,
   textNormalization
 }) {
+  if (!ponyfill.AudioContext) {
+    console.warn(
+      'botframework-directlinespeech-sdk: This browser does not support Web Audio API. Speech support is disabled.'
+    );
+
+    return () => ({});
+  }
+
   return () => {
     const { SpeechGrammarList, SpeechRecognition } = createSpeechRecognitionPonyfillFromRecognizer({
       audioConfig,