+++ samples/13.b.customization-smart-display/src/hooks/useLastBotActivity.js
@@ -1,0 +1,14 @@
+import { hooks } from 'botframework-webchat';
+
+const { useActivities } = hooks;
+
+export default function useLastBotActivity() {
+  const [activities] = useActivities();
+
+  return [
+    activities
+      .slice()
+      .reverse()
+      .find(({ from: { role }, type }) => role === 'bot' && type === 'message')
+  ];
+}